chore: switching bit_index to u8

+++ engine/src/board/bitboard.rs
@@ -6,7 +6,7 @@ pub const BITBOARD_WIDTH: u8 = 8;
 pub const BITBOARD_HEIGHT: u8 = 8;
 
 /// Returns a bitmask as a Bitboard instance.
-fn get_bitboard_for_index(bit_index: u64) -> Bitboard {
+fn get_bitboard_for_index(bit_index: u8) -> Bitboard {
     return Bitboard(get_bitmask_for_index(bit_index))
 }
 
@@ -25,12 +25,12 @@ impl Bitboard {
     }
 
     /// Returns true if bitboard has a bit at given position, false otherwise.
-    pub fn is_set_at(self, bit_index: u64) -> bool {
+    pub fn is_set_at(self, bit_index: u8) -> bool {
         return (self & get_bitboard_for_index(bit_index)) != Bitboard(0);
     }
 
     /// Sets a bit in Bitboard to 1 ato given position.
-    pub fn set_at(&mut self, bit_index: u64) {
+    pub fn set_at(&mut self, bit_index: u8) {
         *self |= get_bitboard_for_index(bit_index);
     }
 }
